requirements.txt
django
djangorestframework
colorlog